requirements.txt
cachetools==4.2.4; python_version >= "3.6" and python_version < "4.0"
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.6; python_full_version >= "3.6.0" and python_version >= "3.6"
environs==9.3.4; python_version >= "3.6"
google-api-core==2.1.0; python_version >= "3.6"
google-api-python-client==2.24.0; python_version >= "3.6"
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6; python_version >= "3.6"
google-auth==2.2.0; python_version >= "3.6"
googleapis-common-protos==1.53.0; python_version >= "3.6"
httplib2==0.20.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
idna==3.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
marshmallow==3.13.0; python_version >= "3.6"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
protobuf==3.18.1; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
python-dotenv==0.19.1; python_version >= "3.6"
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
rsa==4.7.2; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
uritemplate==3.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
